# reduce iterator facade complexity