--- a/power_grid_model_c/power_grid_model/include/power_grid_model/common/iterator_facade.hpp
+++ b/power_grid_model_c/power_grid_model/include/power_grid_model/common/iterator_facade.hpp
@@ -32,22 +32,26 @@ template <class Impl, typename ValueType, std::integral DifferenceType> class It
     constexpr auto operator->() -> decltype(auto) { return &(*(*this)); }
 
     friend constexpr bool operator==(IteratorFacade const& first, IteratorFacade const& second) {
-        return first.equal(second);
+        return (first <=> second) == std::strong_ordering::equivalent;
     }
-    friend constexpr std::strong_ordering operator<=>(IteratorFacade const& first, IteratorFacade const& second)
-        requires requires(const_iterator it) {
-            { it.three_way_compare(it) } -> std::convertible_to<std::strong_ordering>;
-        }
-    {
+    friend constexpr std::strong_ordering operator<=>(IteratorFacade const& first, IteratorFacade const& second) {
         return first.three_way_compare(second);
     }
 
     constexpr auto operator++() -> iterator& {
-        static_cast<iterator*>(this)->increment();
+        if constexpr (requires(iterator it) { it.increment(); }) {
+            static_cast<iterator*>(this)->increment();
+        } else {
+            static_cast<iterator*>(this)->advance(1);
+        }
         return *static_cast<iterator*>(this);
     }
     constexpr auto operator--() -> iterator& {
-        static_cast<iterator*>(this)->decrement();
+        if constexpr (requires(iterator it) { it.decrement(); }) {
+            static_cast<iterator*>(this)->decrement();
+        } else {
+            static_cast<iterator*>(this)->advance(-1);
+        }
         return *static_cast<iterator*>(this);
     }
     constexpr auto operator++(std::integral auto /*idx*/) -> iterator {
@@ -81,10 +85,6 @@ template <class Impl, typename ValueType, std::integral DifferenceType> class It
 
   private:
     // overloads for public bidirectional exposure (difference between MSVC and ClangCL)
-    constexpr bool equal(IteratorFacade const& other) const {
-        return static_cast<std::add_lvalue_reference_t<const_iterator>>(*this).equal(
-            static_cast<std::add_lvalue_reference_t<const_iterator>>(other));
-    }
     constexpr std::strong_ordering three_way_compare(IteratorFacade const& other) const {
         return static_cast<std::add_lvalue_reference_t<const_iterator>>(*this).three_way_compare(
             static_cast<std::add_lvalue_reference_t<const_iterator>>(other));