small refactor after today's mini knowledge sharing session

* [x] TODO: verify no significant performance overhead (do-not-merge label) -> resolved cfr. https://github.com/PowerGridModel/power-grid-model/pull/959#issuecomment-2812613666