@mgovers 2025-04-17
"Benchmark" from CI

run on `main`: https://github.com/PowerGridModel/power-grid-model/actions/runs/14492495556/job/40652046791
run on this branch: https://github.com/PowerGridModel/power-grid-model/actions/runs/14514047165/job/40719016495?pr=959

| job                                                             | `main` | this branch |
| --------------------------------------------------------------- | ------ | ----------- |
| build-cpp-test-linux (debug, gcc)                               | 86.06s | 85.71s      |
| build-cpp-test-linux (debug, clang)                             | 58.76s | 58.53s      |
| build-cpp-test-linux (release, gcc)                             |  9.83s |  9.92s      |
| build-cpp-test-linux (release, clang)                           |  7.40s |  6.79s      |
| build-cpp-test-windows (debug, msvc)                            | 69.62s | 70.16s      |
| build-cpp-test-windows (debug, clang-cl)                        | 31.40s | 31.66s      |
| build-cpp-test-windows (release, msvc)                          |  3.08s |  3.16s      |
| build-cpp-test-windows (release, clang-cl)                      |  2.62s |  2.68s      |
| build-cpp-test-macos (debug)                                    | 41.57s | 41.51s      |
| build-cpp-test-macos (release)                                  | 11.57s | 10.76s      |
| Build and test in Conda (ubuntu)                                | 12.19s | 12.49s      |
| Build and test in Conda (windows)                               | 18.99s | 17.85s      |
| build-and-test-python (linux-x64 - cp311-manylinux_x86_64)      | 24.73s | 24.09s      |
| build-and-test-python (linux-x64 - cp312-manylinux_x86_64)      | 25.44s | 24.42s      |
| build-and-test-python (linux-x64 - cp313-manylinux_x86_64)      | 34.12s | 26.01s      |
| build-and-test-python (linux-x64 - cp311-musllinux_x86_64)      | 27.92s | 27.31s      |
| build-and-test-python (linux-x64 - cp312-musllinux_x86_64)      | 29.03s | 28.42s      |
| build-and-test-python (linux-x64 - cp313-musllinux_x86_64)      | 31.09s | 30.02s      |
| build-and-test-python (linux-aarch64 - cp311-manylinux_aarch64) | 20.65s | 21.12s      |
| build-and-test-python (linux-aarch64 - cp312-manylinux_aarch64) | 21.63s | 20.85s      |
| build-and-test-python (linux-aarch64 - cp313-manylinux_aarch64) | 21.47s | 21.95s      |
| build-and-test-python (macos - cp311-macosx_x86_64)             | 33.46s | 33.33s      |
| build-and-test-python (macos - cp311-macosx_arm64)              | 12.03s | 10.59s      |
| build-and-test-python (macos - cp312-macosx_x86_64)             | 42.42s | 39.10s      |
| build-and-test-python (macos - cp312-macosx_arm64)              | 12.02s | 10.99s      |
| build-and-test-python (macos - cp313-macosx_x86_64)             | 44.03s | 37.23s      |
| build-and-test-python (macos - cp313-macosx_arm64)              | 13.33s | 10.76s      |
| build-and-test-python (windows - cp311-win_amd64)               | 31.36s | 32.63s      |
| build-and-test-python (windows - cp312-win_amd64)               | 30.76s | 33.09s      |
| build-and-test-python (windows - cp313-win_amd64)               | 30.92s | 32.72s      |

Seems like these are indeed similar (within margin; sometimes `main` is faster, other times this branch)